+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/DBAnnotationDriver.java
@@ -1,0 +1,14 @@
+package de.charite.compbio.jannovar.vardbs.base;
+
+/**
+ * Interface for annotation drivers by variant databases.
+ * 
+ * @author Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>
+ */
+public interface DBAnnotationDriver {
+
+	/** @return The {@link VCFHeaderExtender} to use.
+	 */
+	VCFHeaderExtender constructVCFHeaderExtender();
+	
+}