+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/VCFHeaderExtender.java
@@ -1,0 +1,33 @@
+package de.charite.compbio.jannovar.vardbs.base;
+
+import htsjdk.variant.vcf.VCFHeader;
+
+/**
+ * Extend {@link VCFHeader} object with headers for a given database
+ * 
+ * @author Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>
+ */
+public abstract class VCFHeaderExtender {
+
+	/**
+	 * @return Default prefix to use for header entries
+	 */
+	public abstract String getDefaultPrefix();
+
+	/**
+	 * Add header entries with a given prefix.
+	 * 
+	 * @param header
+	 *            The {@link VCFHeader} to extend.
+	 * @param prefix
+	 *            A String prefix to prepend to header record names.
+	 */
+	public abstract void addHeaders(VCFHeader header, String prefix);
+
+	/** Add headers with default prefix
+	 */
+	public void addHeaders(VCFHeader header) {
+		addHeaders(header, getDefaultPrefix());
+	}
+
+}