+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## develop
 
+### jannovar-vardbs
+
+* Initial version of this module, the aim is precise annotation from variant databases.
+
+### jannovar-core
+
 * Adding `altGeneIDs` mapping to `TranscriptModel`, makes data bases backwards incompatible.
 * Rewrite of GFF parsers for RefSeq and ENSEMBL.
 * Bumping HTSJDK to 2.1.0, requiring Java 8 from now on.